requirements.txt
PyQt5>=5.15
paho-mqtt>=1.6
cryptography>=42.0
